Add files via upload
In this digital era, researchers are often faced with the challenge of working with high dimensional dataset. In the worst-case scenario, data mining and machine learning techniques seems to be inefficient against those massive datasets. The gigantic raw dataset including a greater number of features can degrade the performance of machine learning algorithms in terms of accuracy, overfit-ting and complexity as the redundant and irrelevant data do not facilitate the learning process. These issues could be better managed using feature selection technique as it permits the machine learning algorithms to train faster and reduces the complexity of a model, improves the accuracy and reduces overfitting.
The overarching goal of this project is to develop an automated feature selection pipeline by ag-gregating several state-of the art feature selection techniques. The pipeline consists of three types of feature selection techniques: Filter methods, Wrapper methods and Embedded methods. The project implements these methods on the MIMIC-III (Medical Information Mart for Intensive Care) dataset which has the benefit of free access. The database is suitable for the purpose of this project because it is a single-centre database containing information relating to patients admitted to critical care unit at a large territory care hospital which is used for predicting in-hospital mortality after ICU admission. Employing six different classification techniques in the form of Logistic Regression (LR), K-nearest Neighbors (KNN), Decision Tree (DT), Random Forest (RF), Support Vector Machine (SVM) and Artificial Neural Network (ANN), the project compares the performance of the model using several evaluation matrices.